% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{get.s1.sim}
\alias{get.s1.sim}
\title{simulate piecewise constant mean with ARMA noise}
\usage{
get.s1.sim(n, cp, intercept, jump.size, sigma = NULL, arma.mod = NULL)
}
\arguments{
\item{n}{length of time series}

\item{cp}{change points}

\item{intercept}{starting mean}

\item{jump.size}{change(s) in mean}

\item{sigma}{(optional) if ARMA noise is not used}

\item{arma.mod}{(optional) ARMA structure of the noise}
}
\value{
simulated data
}
\description{
simulate piecewise constant mean with ARMA noise
}
\examples{
\donttest{
d = get.s1.sim(n = 200, cp = 100, intercept = 0, jump.size = 1, sigma = 1, arma.mod = c(0.8, 0.2))
}
}
