% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_arma_2.r
\name{not.arma.2}
\alias{not.arma.2}
\title{NOT-ARMA-2}
\usage{
\method{not}{arma.2}(d, contrast, thr.method = "bic")
}
\arguments{
\item{d}{data}

\item{contrast}{A type of the contrast function used in the NOT algorithm. Choice of "pcwsConstMean", "pcwsConstMeanHT", "pcwsLinContMean", "pcwsLinMean", "pcwsQuadMean", "pcwsConstMeanVar". See not::not for more details.}

\item{thr.method}{Threshold method. Currently only BIC is implemented}
}
\value{
a list with fitted NOT model, change points, BIC value and the threshold in NOT
}
\description{
NOT-ARMA-2
}
\examples{
\donttest{
d = get.s1.sim(n = 200, cp = 100, intercept = 0, jump.size = 1, sigma = 1, arma.mod = c(0.8, 0.2))$d
not.arma.2(d, contrast = "pcwsConstMean")
}
}
