% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_arma_2.r
\name{not_arma_2}
\alias{not_arma_2}
\title{NOT-ARMA-2}
\usage{
not_arma_2(
  d,
  contrast,
  thr.method = "bic",
  num.validate = round(length(d) * 0.15)
)
}
\arguments{
\item{d}{data}

\item{contrast}{A type of the contrast function used in the NOT algorithm. Choice of "pcwsConstMean", "pcwsConstMeanHT", "pcwsLinContMean", "pcwsLinMean", "pcwsQuadMean", "pcwsConstMeanVar". See not::not for more details.}

\item{thr.method}{Threshold methods, "bic" for bic, "cv.1" for minimum cv mse, "cv.2" for 1se rule}

\item{num.validate}{optional, number of validation observations}
}
\value{
a list with fitted NOT model, change points, BIC value and the threshold in NOT
}
\description{
NOT-ARMA-2
}
\examples{
\donttest{
d = get.s1.sim(n = 200, cp = 100, intercept = 0, jump.size = 1, sigma = 1, arma.mod = c(0.8, 0.2))$d
not_arma_2(d, contrast = "pcwsConstMean")
}
}
