% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_arma_1.r
\name{predict.not.arma.1.bic}
\alias{predict.not.arma.1.bic}
\title{one step ahead prediction form fitted NOT-ARMA-1 model}
\usage{
\method{predict}{not.arma.1.bic}(object, test.i, ...)
}
\arguments{
\item{object}{fitted model from not.arma.1()}

\item{test.i}{time point to predict}

\item{...}{additional arguments}
}
\value{
one step ahead value
}
\description{
one step ahead prediction form fitted NOT-ARMA-1 model
}
\examples{
\donttest{
d <- get.s1.sim(n = 200, cp = 100, intercept = 0, jump.size = 1, 
sigma = 1, arma.mod = c(0.8, 0.2))$d
mod <- not.arma.1(d[1:199], contrast = "pcwsConstMean")
pred <- predict(mod, 200)
}
}
